finalizing responsiveness
