+++ src/app/globals.css
@@ -24,7 +24,6 @@ body {
 .main {
     position: relative;
     max-width: 1440px;
-    height: 100%;
     margin: 0 auto;
     padding-left: 1rem;
     padding-right: 1rem;
