+++ src/app/page.js
@@ -1,6 +1,6 @@
 import Hero from "@/components/sections/hero";
 import About from "@/components/sections/about";
-import Sponsors from "@/components/sections/sponsors";
+import Sponsors from "@/components/sections/sponsors/sponsors";
 import FAQ from "@/components/sections/faq/faq";
 import Contact from "@/components/sections/contact";
 import Team from "@/components/sections/team/team";
