+++ src/components/sections/about.jsx
@@ -18,7 +18,7 @@ export default function About() {
                 }
             },
             {
-                threshold: 0.2, // Trigger when 10% of the section is visible
+                threshold: 0.1, // Trigger when 10% of the section is visible
             }
         );
