+++ src/components/sections/team/committee.jsx
@@ -3,12 +3,12 @@ import Image from "next/image";
 export default function Committee({ members, title }) {
     return (
         <div className="flex flex-col h-full justify-center items-center lg:pb-[14%]">
-            <h2 className="text-lg lg:text-4xl 3xl:text-5xl font-bold text-center text-white tracking-wide my-6 lg:my-10">
+            <h2 className="text-lg lg:text-4xl 3xl:text-5xl font-bold text-center text-white tracking-wide my-4 md:my-6 lg:my-10">
                 {title}
             </h2>
-            <div className="flex flex-row flex-wrap gap-5 pt-4 items-center justify-center">
+            <div className="flex flex-row flex-wrap gap-5 pt-2 items-center justify-center">
                 {members.map((member, index) => (
-                    <div key={index} className="flex flex-col items-center justify-center w-1/4 lg:w-1/5">
+                    <div key={index} className={`flex flex-col items-center justify-center ${members.length > 4 ? `w-1/4` : `w-1/3`} md:w-1/3 lg:w-1/5`}>
                         <Image
                             src={`/assets/team/${member.img}`}
                             alt={member.name}
