+++ src/components/sections/team/team.jsx
@@ -41,7 +41,7 @@ export default function Team() {
                 }
             },
             {
-                threshold: 0.2, // Trigger when 10% of the section is visible
+                threshold: 0.1, // Trigger when 10% of the section is visible
             }
         );
 
@@ -68,7 +68,7 @@ export default function Team() {
     const dayOf = filterByCommittee(team, "Day-of");
 
     return (
-        <div ref={contentRef} id="team" className="flex flex-col justify-center items-center absolute top-[89%] lg:top-[85.8%] mx-auto w-full h-[9%] lg:h-[13%] float-in-section">
+        <div ref={contentRef} id="team" className="flex flex-col justify-center items-center absolute top-[81.5%] pl-[6%] pr-[3%] pt-2 lg:pt-0 lg:pl-0 lg:pr-0 lg:top-[85.8%] mx-auto w-full h-[17%] lg:h-[13%] float-in-section">
             <Header title="Meet the Team!" />
             <Swiper
                 modules={[Navigation, Pagination, Autoplay]}
@@ -77,7 +77,7 @@ export default function Team() {
                 scrollbar = {{draggable: true }}
                 autoplay={{ delay: 10000, disableOnInteraction: false }}
                 loop={true}
-                className="main w-full mx-10 mt-10 flex items-center justify-center"
+                className="main w-full lg:mx-10 lg:mt-10 flex items-center justify-center"
             >
                 <SwiperSlide>
                     <Committee members={admin} title="Admin" />
